# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.Error:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 43.138.61.49
info:
  contact:
    email: 3264085417@qq.com
    name: big_dust
  description: MucaveAPI
  termsOfService: http://swagger.io/terrms/
  title: Mucave
  version: 1.1.0
paths:
  /login:
    post:
      consumes:
      - application/json
      description: 通过学号密码验证身份
      parameters:
      - description: 学号
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"登录成功，获得token."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":""Invalid username or password}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 登录验证
      tags:
      - login
...
  /user/{id}/user_followers:
    get:
      consumes:
      - application/json
      description: 查询指定id用户的粉丝列表
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定用户的id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到他人的粉丝."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询他人的粉丝失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 用户的粉丝
      tags:
      - user
  /user/{id}/user_following:
    get:
      consumes:
      - application/json
      description: 指定id用户的关注列表
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定用户的id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到他人关注的用户."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询他人关注的用户失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 用户的关注
      tags:
      - user
  /user/{id}/user_msg:
    get:
      consumes:
      - application/json
      description: 查询指定id用户的信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定用户的id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到指定的用户."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询指定用户信息失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 用户信息
      tags:
      - user
  /user/{id}/user_outline:
    get:
      consumes:
      - application/json
      description: 指定id用户的User,动态数,关注数,粉丝数.
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定用户的id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"数据为:user,动态数,关注数,粉丝数."}'
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 指定用户的大致信息
      tags:
      - user
  /user/{id}/user_post:
    get:
      consumes:
      - application/json
      description: 查询到指定id用户的所有帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定用户的id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到他人的帖子."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询他人的帖子失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 指定用户的帖子
      tags:
      - user
  /user/following:
    delete:
      consumes:
      - application/json
      description: 通过关注用户id和被关注用户id删除关注关系
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 被关注者id
        in: query
        name: followed_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"取消关注成功"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"取消关注失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 取消关注
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 通过关注用户id和被关注用户id建立关注关系
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 被关注的用户的id
        in: query
        name: followed_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"关注成功"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"关注失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 关注
      tags:
      - user
  /user/my_comments:
    get:
      consumes:
      - application/json
      description: 查询我的所有评论
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到我的评论."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"我的评论查询失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 我的评论
      tags:
      - user
  /user/my_followers:
    get:
      consumes:
      - application/json
      description: 查询到关注我的用户
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到自己的粉丝."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询自己的粉丝失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 我的粉丝
      tags:
      - user
  /user/my_following:
    get:
      consumes:
      - application/json
      description: 查询我关注的用户
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到自己关注的用户"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询自己关注的用户失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 我的关注
      tags:
      - user
  /user/my_likes:
    get:
      consumes:
      - application/json
      description: 我点赞的所有的帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到我的点赞的帖子."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"我点赞过的帖子查询失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 我的点赞
      tags:
      - user
  /user/my_msg:
    get:
      consumes:
      - application/json
      description: 根据token中的id查我的User体
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到自己的信息."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询自己信息失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 我的信息
      tags:
      - user
    put:
      consumes:
      - application/json
      description: 上传我的新User更新我的信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 名字
        in: formData
        name: name
        required: true
        type: string
      - description: 性别
        in: formData
        name: gender
        required: true
        type: string
      - description: 签名
        in: formData
        name: signature
        required: true
        type: string
      - description: 生日(2006-01-02)
        in: formData
        name: birthday
        required: true
        type: string
      - description: 故乡
        in: formData
        name: hometown
        required: true
        type: string
      - description: 年级
        in: formData
        name: grader
        required: true
        type: string
      - description: 院系
        in: formData
        name: faculties
        required: true
        type: string
      - description: 头像文件
        in: formData
        name: file
        required: true
        type: file
      - description: yes/no(yes:修改头像 no:修改其他信息)
        in: query
        name: avatar_only
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"用户信息修改成功."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"用户信息修改失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 修改我的信息
      tags:
      - user
  /user/my_outline:
    get:
      consumes:
      - application/json
      description: 通过id获取我的大致信息：User,动态数，关注数，粉丝数.
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"数据为:user,动态数,关注数,粉丝数."}'
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 我的大致信息
      tags:
      - user
  /user/my_post:
    get:
      consumes:
      - application/json
      description: 查询到我所有发布过的帖子信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到自己的帖子"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询自己帖子失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 我的帖子
      tags:
      - user
  /user/my_replies:
    get:
      consumes:
      - application/json
      description: 查询我的所有回复
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到我的回复."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"我的回复查询失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 我的回复
      tags:
      - user
  /user/private_msg/{id}:
    get:
      consumes:
      - application/json
      description: 通过id刷新指定用户的向自己发的信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定用户的id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"刷新私信成功"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"刷新私信失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 刷新私信
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 通过id指定发私信的对象进行发私信
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定用户的id
        in: path
        name: id
        required: true
        type: string
      - description: 一个文件
        in: formData
        name: file
        type: file
      - description: yes/no(说明是否附带文件)
        in: query
        name: file_have
        required: true
        type: string
      - description: yes/no(说明是否有文本)
        in: query
        name: content_have
        required: true
        type: string
      - description: 私信的文本内容
        in: formData
        name: content
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"发送私信成功."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"发送私信失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 发私信
      tags:
      - user
  /user/whether_follow:
    get:
      consumes:
      - application/json
      description: 通过用户id查询是否已经关注
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"yes"}'
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 是否关注
      tags:
      - post
schemes:
- http
swagger: "2.0"